[#113257091] Refactor Documents Deletion
Move the documents delete logic to the documents details page
